fix: rename ListNext to useList
